refactor: 로그아웃 훅을 member에서 auth로 이동
