+++ src/features/auth/hooks/useAuthHooks.ts
@@ -1,0 +1,14 @@
+import { useMutation, useQueryClient } from "@tanstack/react-query";
+
+import { authApi } from "@/features/auth/api";
+
+export function useLogout() {
+  const queryClient = useQueryClient();
+
+  return useMutation({
+    mutationFn: () => authApi.logout(),
+    onSuccess: () => {
+      queryClient.clear();
+    },
+  });
+}
